fix(plugin): plugins could not find module same-named as package
